[bq27220] Restructure as a sensor platform
The BQ27220 only exposes sensors, so make it a sensor platform (sensor.py)
with named child sensors, matching peers like ina226, bme280, lc709203f, and
max17043, instead of a top-level hub. Config now uses:

  sensor:
    - platform: bq27220
      battery_level: ...

Removes the AUTO_LOAD workaround; __init__.py now only declares CODEOWNERS.
